+++ changelog.md
@@ -149,3 +149,7 @@ Changelog
   
 1.0.4 - Add Suport
   * Add suport of Combat carousel
+  
+1.0.5 - Add Configuration
+  * Add option of Timing Dialog
+  * Add option of automatic spirit dice charge